update: move custom styles to their own file, refinements
