+++ webapp/IronCalc/src/components/RightDrawer/NamedStyles/edit-named-style.css
@@ -61,7 +61,7 @@
 }
 
 .ic-edit-style-preview {
-  width: 56px;
+  width: 40px;
   height: 32px;
   border-radius: 4px;
   box-shadow: inset 0 0 0 1px
