+++ webapp/IronCalc/src/components/RightDrawer/NamedStyles/named-styles.css
@@ -157,31 +157,53 @@
   position: relative;
 }
 
-.ic-named-styles-pick-list {
+.ic-named-styles-manage-list {
+  flex: 1;
   display: flex;
   flex-direction: column;
-  padding: 8px 0;
 }
 
-.ic-named-styles-pick-item {
-  height: 36px;
-  padding: 0 16px;
+.ic-named-styles-manage-item {
   display: flex;
   align-items: center;
-  background: none;
-  border: none;
-  cursor: pointer;
-  font-size: var(--typography-font-size);
-  font-family: inherit;
-  text-align: left;
+  gap: 12px;
+  padding: 8px 12px 8px 15px;
+  height: 68px;
+  box-sizing: border-box;
+  border-bottom: 1px solid var(--palette-grey-200);
 }
 
-.ic-named-styles-pick-item:hover {
-  background-color: color-mix(
-    in srgb,
-    var(--palette-common-black) 6%,
-    transparent
-  );
+.ic-named-styles-manage-item-preview {
+  width: 40px;
+  height: 36px;
+  border-radius: 4px;
+  box-shadow: inset 0 0 0 1px
+    color-mix(in srgb, var(--palette-common-black) 12%, transparent);
+  flex-shrink: 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: var(--typography-font-size);
+  font-family: var(--typography-font-family);
+  font-weight: 600;
+}
+
+.ic-named-styles-manage-item-name {
+  flex: 1;
+  font-size: var(--typography-font-size);
+  font-family: var(--typography-font-family);
+  color: var(--palette-common-black);
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  font-weight: 600;
+}
+
+.ic-named-styles-manage-item-icons {
+  display: flex;
+  align-items: center;
+  gap: 2px;
+  flex-shrink: 0;
 }
 
 .ic-named-styles-confirm {
