+++ webapp/IronCalc/src/locale/de_de.json
@@ -202,8 +202,7 @@
     "name_already_exists": "Ein Stil mit diesem Namen existiert bereits",
     "new_style": "Neuer Stil",
     "panel_title": "Zellformate",
-    "pick_to_delete": "Stil zum Löschen auswählen",
-    "pick_to_edit": "Stil zum Bearbeiten auswählen",
+
     "update_style": "Stil bearbeiten"
   },
   "navigation": {
