+++ webapp/IronCalc/src/locale/it_it.json
@@ -202,8 +202,7 @@
     "name_already_exists": "Esiste già uno stile con questo nome",
     "new_style": "Nuovo stile",
     "panel_title": "Stili di cella",
-    "pick_to_delete": "Seleziona uno stile da eliminare",
-    "pick_to_edit": "Seleziona uno stile da modificare",
+
     "update_style": "Modifica stile"
   },
   "navigation": {
